Convert expected for comparisons (#19)
* Convert expected values to float before comparison

This converts the expected values to a float before a `<` and `>`
comparison to prevent potential issues around the parsing of integer
values as strings when loading the YAML files. This is particularly of
issue with flat million numbers which are persisted in YAML as 6e+06
which is treated as a string when the YAML is parsed.